pyproject.toml
[tool.poetry]
name = "openai_helper"
version = "0.1.0"
description = ""
authors = ["limdongjin <geniuslim27@gmail.com>"]
readme = "README.md"
packages = [{include = "openai_helper"}]

[tool.poetry.dependencies]
python = "^3.9"
openai = "^0.27.2"
hydra-core = "^1.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
